LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/background-repeat-expected.txt


FAIL Can set 'background-position' to CSS-wide keywords: initial assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: inherit assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: unset assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: revert assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'background-position' to the 'repeat-x' keyword: repeat-x Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'repeat-y' keyword: repeat-y Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'repeat' keyword: repeat Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'space' keyword: space Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'round' keyword: round Bad value for shorthand CSS property
FAIL Can set 'background-position' to the 'no-repeat' keyword: no-repeat Bad value for shorthand CSS property
FAIL Setting 'background-position' to a length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'background-position' to a percent throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'background-position' to a time throws TypeError
PASS Setting 'background-position' to an angle throws TypeError
PASS Setting 'background-position' to a flexible length throws TypeError
FAIL Setting 'background-position' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'background-position' to a URL throws TypeError
PASS Setting 'background-position' to a transform throws TypeError
FAIL 'background-position' does not supported 'space repeat' assert_not_equals: Unsupported value must not be null got disallowed value null
